refactor(application): cleanse individual product Price calculation logic

+++ src/main/java/kitchenpos/dao/JdbcTemplateMenuProductDao.java
@@ -74,6 +74,6 @@ public class JdbcTemplateMenuProductDao implements MenuProductDao {
             resultSet.getLong(KEY_COLUMN_NAME),
             resultSet.getLong("menu_id"),
             resultSet.getLong("product_id"),
-            resultSet.getLong("quantity"));
+            resultSet.getInt("quantity"));
     }
 }
